feat: simple distributor contract

+++ src/SubtensorStorage.sol
@@ -17,11 +17,11 @@ library SubtensorStorage {
     bytes16 internal constant PALLET_PREFIX = 0x658faa385070e074c85bf6b568cf0555;
 
     // Storage item name hashes
-    bytes16 internal constant SUBNET_ALPHA_IN_EMISSION_PREFIX  = 0x1905df3b2516a166b6f9fba54fef1cd8;
+    bytes16 internal constant SUBNET_ALPHA_IN_EMISSION_PREFIX = 0x1905df3b2516a166b6f9fba54fef1cd8;
     bytes16 internal constant SUBNET_ALPHA_OUT_EMISSION_PREFIX = 0x25257fbc5458419b7bc7e8c44c521521;
-    bytes16 internal constant SUBNET_TAO_IN_EMISSION_PREFIX    = 0xdd62ae7237581e8f6a684f1ecae06215;
-    bytes16 internal constant SUBNET_ALPHA_IN_PREFIX           = 0x2ce12f7007574647d692ac7edf8b7a53;
-    bytes16 internal constant SUBNET_ALPHA_OUT_PREFIX          = 0x7837978cc6746112a2c9e680a18cfcb9;
+    bytes16 internal constant SUBNET_TAO_IN_EMISSION_PREFIX = 0xdd62ae7237581e8f6a684f1ecae06215;
+    bytes16 internal constant SUBNET_ALPHA_IN_PREFIX = 0x2ce12f7007574647d692ac7edf8b7a53;
+    bytes16 internal constant SUBNET_ALPHA_OUT_PREFIX = 0x7837978cc6746112a2c9e680a18cfcb9;
 
     // --------------------------------------------------------------------- //
     // INTERNALS                                                             //
@@ -31,18 +31,14 @@ library SubtensorStorage {
      * SCALE-encode a `u16` (little-endian).
      */
     function _encodeNetuid(uint16 netuid) private pure returns (bytes2) {
-        return bytes2(netuid);                  // solidity stores uint16 LE in bytesN
+        return bytes2(netuid); // solidity stores uint16 LE in bytesN
     }
 
     /**
      * Build the raw storage key:
      *    key = twox128(pallet) ++ twox128(item) ++ SCALE(key-arg)
      */
-    function _buildStorageKey(bytes16 itemHash, uint16 netuid)
-        private
-        pure
-        returns (bytes memory)
-    {
+    function _buildStorageKey(bytes16 itemHash, uint16 netuid) private pure returns (bytes memory) {
         // 16 + 16 + 2 = 34 bytes
         return abi.encodePacked(PALLET_PREFIX, itemHash, _encodeNetuid(netuid));
     }
@@ -58,18 +54,21 @@ library SubtensorStorage {
         bool success;
 
         assembly {
-            success := staticcall(
-                gas(),                  // gas
-                PRECOMPILE_ADDRESS,     // address of precompile
-                add(fullStorageKey, 0x20), // input ptr (skip length)
-                mload(fullStorageKey),      // input size
-                add(out, 0x20),         // output ptr (skip length)
-                8                       // output size (request up to 8 bytes for u64)
-            )
+            success :=
+                staticcall(
+                    gas(), // gas
+                    PRECOMPILE_ADDRESS, // address of precompile
+                    add(fullStorageKey, 0x20), // input ptr (skip length)
+                    mload(fullStorageKey), // input size
+                    add(out, 0x20), // output ptr (skip length)
+                    8 // output size (request up to 8 bytes for u64)
+                )
         }
 
         uint256 returnedDataSize;
-        assembly { returnedDataSize := returndatasize() }
+        assembly {
+            returnedDataSize := returndatasize()
+        }
 
         if (!success || returnedDataSize == 0) {
             return 0; // Call failed or no data (e.g., key not found)
@@ -80,10 +79,12 @@ library SubtensorStorage {
         uint64 tempValue = 0;
         uint256 maxBytesToRead = returnedDataSize < 8 ? returnedDataSize : 8;
 
-        for (uint i = 0; i < maxBytesToRead; ) {
+        for (uint256 i = 0; i < maxBytesToRead;) {
             uint8 byteVal;
             // solhint-disable-next-line no-inline-assembly
-            assembly { byteVal := byte(0, mload(add(add(out, 0x20), i))) }
+            assembly {
+                byteVal := byte(0, mload(add(add(out, 0x20), i)))
+            }
             tempValue |= uint64(byteVal) << uint64(i * 8);
             // solhint-disable-next-line no-plusplus
             i++;
@@ -114,4 +115,4 @@ library SubtensorStorage {
     function subnetAlphaOut(uint16 netuid) internal view returns (uint64) {
         return queryPrecompile(_buildStorageKey(SUBNET_ALPHA_OUT_PREFIX, netuid));
     }
-} 
+}
